fix(web): prefer html for crawlers
